src/requirements/prd.txt
aiosqlite==0.17.0
alembic==1.6.5 # from 1.5.8
asyncpg==0.23.0 # from 0.22.0
databases==0.4.3
devsetgo-lib==0.5.0
fastapi-login
fastapi[all]==0.65.2 # from 0.64.0
gunicorn==20.1.0
httpx==0.18.1
loguru==0.5.3
orjson==3.5.3
passlib[bcrypt]==1.7.4
psutil==5.8.0
py-cpuinfo==8.0.0
python-dotenv==0.17.1
python-jose==3.3.0
python-jose==3.3.0
python-multipart==0.0.5
pyyaml==5.4.1
silly==1.1.0
starlette-exporter==0.8.2
unsync==1.3.2
xmltodict==0.12.0
